Use round-robin as the default dispatch handler.
